添加自定义用户授权token的验证及其注解

+++ newenergy-db/src/main/java/newenergy/db/service/PlotFactorService.java
@@ -108,7 +108,7 @@ public class PlotFactorService implements Searchable<ApplyFactor, ApplyFactorPre
         /**
          * TODO 等待修改类型
          */
-        BigDecimal originFactor = BigDecimal.valueOf(res.getPlotFactor());
+        BigDecimal originFactor = res.getPlotFactor();
         applyFactor.setOriginFactor(originFactor);
         return repository.saveAndFlush(applyFactor)==null?ResultConstant.ERR:ResultConstant.OK;
     }
